# openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Issues Gateway
  version: 1.0.0
  summary: A contract-first HTTP gateway over the GitHub Issues REST API.
  description: |
    Wraps the GitHub Issues API for one configured repository
    (`GITHUB_OWNER/GITHUB_REPO`) behind a stable, normalised HTTP interface,
    and receives signed GitHub webhooks.

    ### On "delete"
    The GitHub REST API has no delete-issue operation. The **D** of CRUD is
    therefore a state transition: `PATCH /issues/{number}` with
    `{"state": "closed"}`. `DELETE /issues/{number}` is offered as an alias
    for that close and returns the closed issue rather than `204`, so no
    caller is misled into thinking data was destroyed.

    ### Authentication
    Callers of *this* API do not present a GitHub token. The service holds a
    fine-grained PAT server-side (`GITHUB_TOKEN`) and uses it for every
    upstream call. The `githubToken` security scheme below documents that
    server-to-GitHub credential.

    Webhook deliveries authenticate differently: with an HMAC SHA-256
    signature in `X-Hub-Signature-256`, verified against `WEBHOOK_SECRET`.

    ### Errors
    Every non-2xx response uses one envelope, the `Error` schema, so a client
    needs exactly one error branch.
  contact:
    name: CMPE 272 — Issues Gateway
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8080"
...
security:
  - githubToken: []
paths:
...
  /webhook:
    post:
      tags: [webhooks]
      summary: Receive a GitHub webhook delivery
      description: |
        Verifies `X-Hub-Signature-256` (HMAC SHA-256 over the **raw** body,
        compared in constant time) before parsing or trusting anything.

        Accepts `issues`, `issue_comment` and `ping`. Deliveries are stored
        under the composite key `(X-GitHub-Delivery, action)` with
        `INSERT OR IGNORE`, so a GitHub retry or a manual redelivery is a
        no-op rather than a duplicate. The `204` is written before any
        follow-on work runs, keeping the ack well inside GitHub's delivery
        timeout.
      operationId: receiveWebhook
      security:
        - webhookSignature: []
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          description: Event name.
          schema:
            type: string
            enum: [issues, issue_comment, ping]
        - name: X-GitHub-Delivery
          in: header
          required: true
          description: Delivery UUID; the primary component of the dedupe key.
          schema:
            type: string
            format: uuid
        - name: X-Hub-Signature-256
          in: header
          required: true
          description: "`sha256=` followed by the hex HMAC of the raw body."
          schema:
            type: string
            pattern: "^sha256=[0-9a-f]{64}$"
      requestBody:
        required: true
        description: The raw GitHub event payload.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
            examples:
              issueOpened:
                summary: issues / opened
                value:
                  action: opened
                  issue:
                    number: 42
                    title: Login button is misaligned on mobile
                    state: open
                  repository:
                    full_name: andrewbond/cmpe272-issues-gw
                  sender:
                    login: andrewbond
              commentCreated:
                summary: issue_comment / created
                value:
                  action: created
                  issue:
                    number: 42
                  comment:
                    id: 2938475610
                    body: Reproduced on iPhone 14.
                  sender:
                    login: andrewbond
              ping:
                summary: ping sent when the webhook is saved
                value:
                  zen: Non-blocking is better than blocking.
                  hook_id: 55512345
      responses:
        "204":
          description: |
            Delivery accepted, or recognised as a duplicate and ignored. Both
            cases return 204 so GitHub's retry logic settles either way.
        "400":
          description: Unknown event, unknown action, or malformed JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unsupportedEvent:
                  value:
                    error:
                      code: unsupported_event
                      message: Unsupported event type 'push'.
                unsupportedAction:
                  value:
                    error:
                      code: unsupported_action
                      message: Unsupported action 'exploded' for event 'issues'.
        "401":
          description: Missing, malformed or invalid signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                badSignature:
                  value:
                    error:
                      code: invalid_signature
                      message: Webhook signature verification failed.
  /events:
    get:
      tags: [webhooks]
      summary: List recently processed webhook deliveries
      description: A debugging aid — the newest deliveries this service stored.
      operationId: listEvents
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum deliveries to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: Stored deliveries, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEvent"
              examples:
                recent:
                  value:
                    - id: 8f2c1e40-9b3a-11f0-8d1e-2f9a7c4b1a55
                      event: issues
                      action: opened
                      issue_number: 42
                      sender: andrewbond
                      timestamp: "2026-09-08T17:04:12+00:00"
                    - id: 91b0aa22-9b3a-11f0-9c77-1d55e2b0f331
                      event: issue_comment
                      action: created
                      issue_number: 42
                      sender: andrewbond
                      timestamp: "2026-09-08T18:22:04+00:00"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub fine-grained PAT
      description: |
        The credential this **service** presents to GitHub, supplied through
        the `GITHUB_TOKEN` environment variable and never exposed to callers
        of this API. Documented here because every `/issues*` operation
        depends on it: without it those routes return `401`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        HMAC SHA-256 of the raw request body, keyed with `WEBHOOK_SECRET`,
        rendered as `sha256=<hex>`. Verified with a constant-time compare.
...
  schemas:
...
    WebhookEvent:
      type: object
      title: WebhookEvent
      description: A stored webhook delivery.
      required: [id, event, timestamp]
      properties:
        id:
          type: string
          description: The `X-GitHub-Delivery` UUID.
        event:
          type: string
          enum: [issues, issue_comment]
        action:
          type: [string, "null"]
          examples: [opened, closed, reopened, created]
        issue_number:
          type: [integer, "null"]
        sender:
          type: [string, "null"]
        timestamp:
          type: string
          format: date-time
          description: When this service received the delivery (UTC).
    WebhookPayload:
      type: object
      title: WebhookPayload
      description: |
        GitHub's event body. Only the fields this service reads are described;
        GitHub sends many more and they are stored verbatim.
      properties:
        action:
          type: string
        issue:
          type: object
          properties:
            number:
              type: integer
            title:
              type: string
            state:
              type: string
        comment:
          type: object
          properties:
            id:
              type: integer
            body:
              type: string
        repository:
          type: object
          properties:
            full_name:
              type: string
        sender:
          type: object
          properties:
            login:
              type: string
        zen:
          type: string
          description: Present only on `ping`.
        hook_id:
          type: integer
          description: Present only on `ping`.
    ErrorDetail:
      type: object
      title: ErrorDetail
      required: [message]
      properties:
        field:
          type: [string, "null"]
          description: Dotted path of the offending field, when one applies.
        message:
          type: string
    Error:
      type: object
      title: Error
      description: The single error envelope used by every non-2xx response.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier.
              enum:
                - invalid_request
                - unauthorized
                - invalid_signature
                - missing_event_header
                - unsupported_event
                - unsupported_action
                - malformed_payload
                - forbidden
                - not_found
                - rate_limited
                - upstream_error
                - upstream_unavailable
                - http_error
                - internal_error
            message:
              type: string
              description: Human-readable summary. Never contains a token or signature.
            details:
              type: array
              items:
                $ref: "#/components/schemas/ErrorDetail"
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingTitle:
              summary: Required field absent
              value:
                error:
                  code: invalid_request
                  message: Request failed validation.
                  details:
                    - field: title
                      message: Field required
            invalidState:
              summary: Value outside the allowed enum
              value:
                error:
                  code: invalid_request
                  message: Request failed validation.
                  details:
                    - field: state
                      message: "Input should be 'open', 'closed' or 'all'"
            emptyPatch:
              summary: Patch with no fields
              value:
                error:
                  code: invalid_request
                  message: Request failed validation.
                  details:
                    - field: null
                      message: >-
                        Value error, at least one of 'title', 'body' or 'state'
                        must be provided
...